Move plotting functions in sensor_functions to mpl_function.
Module sensor_functions in reco was littered with plotting functions
that belong now to mpl_functions in icaro. This commit moves all those
functions to their proper place.
